// src/app/models/conferencistas.interface.ts
export interface conferencistas{
    id?: string;
    nombre: string;
    tipo: string;
    nomConferencia: string;
    desConferencia: string;
    datosPonente: string;
    datosPonente1?: string;
    datosPonente2?: string;
    datosPonente3?: string;
    fecha: string;
    fecha1?: string;
    fecha2?: string;
    link: string;
    link1?: string;
    link2?: string;
    img: string;
    idx?: number;
    completed?: boolean;
}